pyproject.toml
[project]
name = "crabot"
version = "0.1.0"
description = "random basic discord bot"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "asyncio>=4.0.0",
    "google-adk>=1.27.2",
    "pydantic>=2.12.5",
]
